Validate article id and user for /article DELETE
Forbids attempts of deleting articles the request user is not an owner
of.
Validates the article id passed is a valid mongodb id, and exists in the
articles db.
Extracts the handling of Flask's g object to get the current Firebase
user to a separate function, so it can be mocked during tests. Every
test should probably mock this function, or make new ones when handling
the g object, because it's apparently not possible to do it directly on
the test cases.