AND/OR filters for mongo models
Enhances the get_many Model method to support a more complex (but still
basic) filtering, with multiple parameters chained together in an AND
sequence, and list arguments for parameters being treated as an OR
filter of each element.

This implementation forces the model's schema to be single typed, and
forbids the usage of union types. These filters are supposed to be used
with a GET HTTP request querystring, where all values come as a string.
If we use union types we just don't know which type to cast to, and we'd
have to iterate over every single type until one magically fits our
variable. The easiest solution was to remove this flexibility and force
a single type upon each variable. Optional types are handled specially,
and so is the int-float case. The Article model's schema is changed from
using Real (abstract base class which int and float implement) to plain
float to fit this.